+++ Collectors/MarkSweepCollector.cpp
@@ -40,9 +40,9 @@ void MarkSweepCollector::collect(int reason) {
 }
 
 void MarkSweepCollector::checkWatermark() {
-	int size = myAllocator->getRegionSize();
-	int free = myAllocator->getFreeSize();
-	int ratio = 100 - (100 * free / size);
+	size_t size = myAllocator->getRegionSize();
+	size_t free = myAllocator->getFreeSize();
+	size_t ratio = 100 - (100 * free / size);
 	if (ratio > myWatermark) {
 		collect((int)reasonHighWatermark);
 	}
